+++ projects/core/src/file/file.ts
@@ -3,4 +3,8 @@ import { Input } from "@angular/core";
 export abstract class NgEzFileBase{
 
     @Input() accept: string;
+
+    protected fileInput: HTMLInputElement;
+
+    protected listener: Function;
 }