+++ projects/core/src/file/index.ts
@@ -1,3 +1,4 @@
 export * from './file-dropzone.directive';
 export * from './file-input.directive';
 export * from './file.module';
+export * from './models';