+++ projects/core/src/file/models.ts
@@ -1,0 +1,5 @@
+export class NgEzFileDropzoneEvent{
+    constructor(private event:NgEzFileDropzoneEventTypes, private files: File[]){}
+}
+
+export type NgEzFileDropzoneEventTypes = 'drop' | 'select';