+++ projects/io/src/assets/code-snippets/core/file-dropzone/app.component.ts
@@ -1,5 +1,5 @@
-import { Component } from "@angular/core";
-import { NgEzByteUtils, NgEzValidators } from "@ngez/core";
+import { Component, ViewChild } from "@angular/core";
+import { NgEzByteUtils, NgEzValidators, NgEzFileDropzoneDirective } from "@ngez/core";
 import { FormBuilder, FormGroup } from "@angular/forms";
 
 @Component({
@@ -7,6 +7,8 @@ import { FormBuilder, FormGroup } from "@angular/forms";
     templateUrl: './app.component.html'
 })
 export class AppComponent {
+
+    @ViewChild(NgEzFileDropzoneDirective) dropzone: NgEzFileDropzoneDirective;
 
     form: FormGroup;
 
@@ -19,6 +21,11 @@ export class AppComponent {
     onRemove(file: File){
         const control = this.form.get('files');
         control.setValue((control.value as File[]).filter(f => file != f))
+    }
+
+    //Programtically open file browser
+    onBrowse(){
+        this.dropzone.browse();
     }
 
 }